Distribution alterations and comments
Exponential needed to be negative exponential and centered around a
double value.
Poisson just needed to be centered around a specified double value and
random needed to be created outside of the method so that it actually
generated random values as the random number objects are generated by
the system clock.